Pass VM ID to `Function:LoadInstance` instead of relying on existing one
This will also come in handy when the function's VM ID will be assigned to
it after a VM for it is loaded from a snapshot (i.e., a new VM is created
rather than an offloaded one is reused). This logic follows the branch when
a snapshot is not available and a VM is booted from scratch.

Signed-off-by: Georgiy Lebedev <lebedev.gk@phystech.edu>