Pass VM ID by value to goroutine for `Function:RemoveInstanceAsync`
Goroutines capture variables by reference, so it is safer to make a copy
(i.e., a local variable) and pass the variable by value.

This will also come in handy when the function's VM ID can change when it
is loaded from a snapshot booted VM rather than an offloaded one.

Signed-off-by: Georgiy Lebedev <lebedev.gk@phystech.edu>